bundle infra for rendering
